+++ src/module/data/actor/_module.mjs
@@ -1,3 +1,3 @@
 export { default as BaseActorModel } from "./base.mjs";
-export { default as CharacterModel } from "./character.mjs";
+export { default as HeroModel } from "./hero.mjs";
 export { default as NPCModel } from "./npc.mjs";