+++ src/module/data/actor/_types.d.ts
@@ -68,13 +68,13 @@ declare module "./base.mjs" {
   }
 }
 
-declare module "./character.mjs" {
+declare module "./hero.mjs" {
   type HeroicResource = {
     value: number;
     label?: string;
   };
 
-  export default interface CharacterModel {
+  export default interface HeroModel {
     recoveries: BarAttribute & {
       bonus: number;
       recoveryValue: number;