+++ src/module/data/item/feature.mjs
@@ -2,7 +2,7 @@ import { systemPath } from "../../constants.mjs";
 import AdvancementModel from "./advancement.mjs";
 
 /**
- * Passive benefits usually granted by other items.
+ * A data model directly representing class and monster features as well as the basis for ancestry traits, perks, and titles.
  */
 export default class FeatureModel extends AdvancementModel {
   /** @inheritdoc */
@@ -10,8 +10,8 @@ export default class FeatureModel extends AdvancementModel {
     return {
       ...super.metadata,
       type: "feature",
-      packOnly: false,
-      detailsPartial: [systemPath("templates/sheets/item/partials/feature.hbs")],
+      // Not currently in use
+      // detailsPartial: [systemPath("templates/sheets/item/partials/feature.hbs")],
     };
   }
 
@@ -27,29 +27,8 @@ export default class FeatureModel extends AdvancementModel {
     const fields = foundry.data.fields;
     const schema = super.defineSchema();
 
-    schema.type = new fields.SchemaField({
-      value: new fields.StringField({ required: true }),
-      subtype: new fields.StringField({ required: true }),
-    });
-
-    // Can be expanded over time for automation
-    schema.prerequisites = new fields.SchemaField({
-      value: new fields.StringField({ required: true }),
-    });
+    // TODO: https://github.com/MetaMorphic-Digital/draw-steel/issues/839 will re-add content here.
 
     return schema;
-  }
-
-  /* -------------------------------------------------- */
-
-  /** @inheritdoc */
-  async getSheetContext(context) {
-    const featureConfig = ds.CONFIG.features;
-
-    context.featureTypes = Object.entries(featureConfig.types).map(([value, entry]) => ({ value, label: entry.label }));
-
-    if (featureConfig.types[this.type.value]?.subtypes) {
-      context.featureSubtypes = Object.entries(featureConfig.types[this.type.value].subtypes).map(([value, { label }]) => ({ value, label }));
-    }
   }
 }